featurizer/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
python-dateutil==2.8.2
numpy==1.26.2
pandas==2.1.4
pytest==7.4.3
 
